Changed blunderbuss reload number from int to float, increased reload time for all forms of blunderbuss reload, increased range of blundrebuss

+++ Source/MI498_UEProject/Weapons/Blunderbuss/Blunderbuss.h
@@ -34,7 +34,7 @@ public:
 	
 	/// How much time it will take to reload the blunderbuss if it is missing two ammo
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
-	int TwoAmmoReloadTime = 1;
+	float TwoAmmoReloadTime = 1;
 	
 	/// How much force is applied to the player when firing the weapon in air during the primary attack
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
